requirements-production.txt
# Production requirements for high-volume eCommerce
# Core Django
Django==5.1.4
asgiref==3.8.1
sqlparse==0.5.3
python-decouple==3.8

# Database (PostgreSQL)
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Caching (Redis)
redis==5.0.1
django-redis==5.4.0

# Background Tasks (Celery)
celery==5.3.4
django-celery-beat==2.5.0

# Performance & Monitoring
sentry-sdk==1.38.0
django-ratelimit==4.1.0
django-cors-headers==4.3.1

# Cloud Storage
django-storages==1.14.2
boto3==1.34.0

# Forms
django-countries==7.6.1
django-crispy-forms==2.3
crispy-bootstrap4==2024.10

# Image Processing
Pillow==11.1.0

# PDF Generation
weasyprint==63.1
reportlab==4.2.5

# API
djangorestframework==3.15.2
drf-spectacular==0.27.0

# Search
elasticsearch==8.11.0
django-elasticsearch-dsl==8.0.0

# Additional utilities
requests==2.32.3
python-dateutil==2.8.2